refactor iterator, multiple fixes, make tests pass
Fixes #43 #40 and #3. Unfortunately not easy to split out in separate commits.

1. maxogden/level.js#43 "snapshots seem not to work"

The reason snapshots didn't work in the test, is because the iterator was only
opened on the first call to _next. The new version opens the iterator right
away in the constructor (if there is no KeyRange error).

2. maxogden/level.js#40 "abstract-leveldown tests aren't all passing anymore"

Partially fixed by merging PR #42 and further fixed by this commit.

3. maxogden/level.js#3 "stop batch reads after limit is reached"

The solution is stop calling cursor.continue() once the limit is reached. The
transaction will automatically timeout as prescribed by the spec. idb-wrapper
can't be used for this because limit is only respected if autoContinue is true.

All 563 tests pass on:
* Safari 9.0.3
* Firefox 45.0b2
* Firefox 46.0a2
* Iridium 44.1 (Chrome/44.0.2403.157)